feat: 카드 등록 페이지에 필요한 컴포넌트 구조 개선
