+++ src/components/CardRegister/CardCompany.tsx
